// apps/web/src/lib/api/generated/model/receivedPayment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */
import type { ClientRef } from './clientRef';
import type { PaymentAllocationOut } from './paymentAllocationOut';

export interface ReceivedPayment {
  allocations: PaymentAllocationOut[];
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  amount: string;
  client: ClientRef;
  created_at: string;
  currency: string;
  id: string;
  method: string;
  notes: string | null;
  number: string;
  received_on: string;
  reference: string | null;
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  unallocated: string;
  void_reason: string | null;
  voided_at: string | null;
}